Remove old longrun/test_remote_DB.py module
